vision types
